This is the initial code for the a GitHub Action that will setup a native installation of SQLite.
These set of changes include suggestions that were taken from the
javascript-action and setup-node repositories.

Thou those inputs have been included from those two repositories.  This
doesn't imply that these are going to be used going forward.  These
could be replaced and/or updated in the future.  I just wanted to thank
them for this.