Added the output sqlite-bin that will be set to the directory in which that sqlite executable is located.  This can be used to copy the installed version of SQLite such that it can be packaged with a distribution.
As part of the addition of the sqlite-bin output value.  A new test was
added to test that the different output values have been assigned
correctly for different combinations of the setup-sqlite action.

This update required that the version of this setup-sqlite action be
bumped to version 2.

This work required that the README file needed to be updated to reflect
these changes.  As part of that work, some of the documentation was
updated to make it clearer how this action can be used.