The prior GitHub REST APIs used to get the tags information has changed.
The latest version of GitHub REST APIs have changed with respect to how to
get information about a repositories tags.  This required the code to be
updated to use the newer version of the rest api.

While the above change works with retrieveing the version information
for SQLite from GitHub.  There are still test failures that are not
related to the above changes.  Those tests will be fixed within a
different commit.

This is related to issue #6.