The SQLite bundle names have changed starting from version 3.44.0.
These set of changes fixes an issue with the fact that the different
SQLite bundle names have changed since version 3.44.0.  Thus, these changes
insures that we correctly generate the bundle file name depending on the
particular version of SQLite.  Thus, the changes will work with versions
prior to 3.44.0 and with versions 3.44.0 or greater.

As part of the above changes, these changes required that some tests
were added that will test against version 3.44.0 and the prior version
3.43.2.

These changes are associated with issue #5.