Added code to be able to handle GitHub REST API rate limits.
These set of changes adds code that will manage the case where we've
reach the rate limit when using the GitHub REST APIs.  The additional
code will use the returned information to determine how long before
trying the REST API call again.  It will then wait for that amount of
time and try again.  The code also will only retry for a total of 3
times.  This will insure that we won't be retrying forever.

As part of the above addition, a test was added that will reproduce a
rate limit issue and still be able to successfully process the request.

The above changes are related to issue #7.