Added the sqlite-retry-count input option.
This set of changes adds the sqlite-retry-count input parameter that is
used to set the maximum number of reties the setup sqlite action will
perform when using the github rest api calls.  This input is only useful
whenever the sqlite version/year has not been set.  These types of
combinations will require the use of the github rest api calls to get
the year information pertaining to the requested sqlite version.  Those
calls can generate a rate limit response that is used to determine how
long one needs to wait before making the githut rest call.  While these
retries can succeed, there is always a chance that they will fail again
because of concurrent github rest calls.  While the default retry count
of 3 should suffice for most cases.  The ability to increase this retry
count can be of benefit where this action is used a lot concurrently.

As part of these changes, several unit tests were added that checks that
the retry count is set correctly.  The workflow tests that don't use the
sqlite version/year combination updates the retry count to 10.  This
should allow all tests to complete successfully.

This work is part of issue #9.