This set of changes resolves an issue with the three group of tests.
   - retry count
   - setup monthly
   - setup

These tests use the same directory structure for the runners temp and
tool cache directory.  This was causing the workflow runs to fail
because of one or more of the above tests failed because of the
premature deletion of the runners temp and tool cache directories.

The runner temp and tool cache directories are now unique for each of
the above sets of tests.  This will insure that the runners temp and
tool cache directories will not be deleted prematurely during the
execution of those tests.