The retry count implementation incorrectly performed the validation check.
This simple change corrects a minor issue with the way that we check if
the sqlite-retry-count input variable is a correct integer value.  The
prior implementation performed the check in an incorrect order.  This
change reverses the order checking and correctly determines if the
passed input value is an acceptable integer value.

This is part of issue #9.