This set of changes fixes an issue with the retry count unit tests.
These tests were not actually testing the retry count feature properly.
The issue is that the variable that we checked the results against were
not producing the expected result because the variable is a constant.
The test in its original form incorrectly determined the expected result
and incorrectly stated that the test passed.  Upon fixing the function that
determines the expected result.  The retry count tests started to fail.
While the test failures themselves didn't reveal that the implementation
of this feature was incorrect.  It did show that the tests were
implemented incorrectly.  Since the variable the tests were comparing
against was a constant.  Each of these test runs required a single test
file that tested that one test combination.  Thus, the creation of the
retry.count.js file that is referenced by the retry.count.*.test.js
tests.  Upon creating those files all of the retry count tests generated
the correct test results.

As part of the above work.  The default test target was updated such
that it will run all of the tests except for the monthly tests.  Also
there was the addition of the test-monthly target that is used by the
monthly schedule workflow.  This was done to make it easier to execute
the different tests.

Lastly, made some minor changes to the message being displayed when
setting the retry count within the action initialization process.