The following set of changes contains fixes to the test run for the bimonthly tests.  The prior test command was not finding the bimonthly tests.  The issue was with the jest command line option.  This has been corrected and the bimonthly tests are being executed when calling the test-bimonthly command.
As part of the above, the workflow file was renamed to reflect that
these tests are executed bimonthly instead of monthly.  As part of this
changes the README file was also updated to reflect this changes.

Finally, the version was bumped up to version 1.2.1 in preparation to
the next release of this action.